requirements.txt
whisperx
torch
torchaudio
pyannote.audio
python-dotenv
